# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GSNpeeps API
  version: 0.7.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    Kontrak HTTP GSNpeeps untuk Employee Database, dashboard HR, absensi,
    approval, notifikasi, dan administrasi akses.

    API menggunakan field JSON `snake_case`, Bearer JWT berumur 8 jam, dan
    cross-check sesi aktif di Redis. Tidak ada refresh-token endpoint.
  contact:
    name: GSNpeeps IT Liaison
servers:
  - url: http://{host}:{port}
    description: Development
    variables:
      host:
        default: localhost
      port:
        default: "8080"
  - url: https://{host}
    description: Production
    variables:
      host:
        default: api.janjikupadamu.id
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/auth/me/foto:
    put:
      operationId: updateMyPhoto
      summary: Perbarui foto profil sendiri
      description: |
        Tersedia untuk seluruh role terautentikasi, termasuk Top Management yang tidak
        memiliki Profil Saya, karena navbar menampilkan foto untuk seluruh role. File
        diteruskan backend ke Nextcloud; database hanya menyimpan locator (D-037).
      tags: [Authentication]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PhotoUploadRequest"
      responses:
        "200":
          description: Foto profil berhasil diperbarui.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoUpdateResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "413": { $ref: "#/components/responses/FileTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedFileType" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/v1/master/departemen:
    get:
      operationId: listDepartments
      summary: Ambil master departemen
      description: Dropdown organisasi read-only untuk seluruh role terautentikasi.
      tags: [Master Organization]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Daftar departemen.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/master/jabatan:
    get:
      operationId: listPositions
      summary: Ambil master jabatan
      description: Dropdown jabatan read-only untuk seluruh role terautentikasi.
      tags: [Master Organization]
      security:
        - BearerAuth: []
      parameters:
        - name: department_id
          in: query
          required: false
          description: Filter jabatan berdasarkan departemen.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Daftar jabatan.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token berlaku 8 jam. Backend juga memverifikasi session aktif di
        Redis pada setiap request terproteksi.
...
  responses:
    BadRequest:
      description: Request atau parameter tidak valid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            invalidRequest:
              value:
                success: false
                error: { code: INVALID_REQUEST, message: Request tidak valid }
    Unauthorized:
      description: Kredensial tidak ada, tidak valid, atau kedaluwarsa.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            unauthorized:
              value:
                success: false
                error: { code: UNAUTHORIZED, message: Autentikasi diperlukan }
...
    ValidationError:
      description: Validasi bisnis atau field gagal.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            validation:
              value:
                success: false
                error:
                  code: VALIDATION_ERROR
                  message: Validasi gagal
                  fields: { tanggal_mulai: Tanggal mulai wajib diisi }
    FileTooLarge:
      description: File melebihi batas 5 MB.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            fileTooLarge:
              value:
                success: false
                error: { code: FILE_TOO_LARGE, message: Ukuran file maksimal 5 MB }
    UnsupportedFileType:
      description: Tipe file tidak didukung.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            unsupported:
              value:
                success: false
                error: { code: UNSUPPORTED_FILE_TYPE, message: Tipe file tidak didukung }
    TooManyRequests:
      description: Rate limit terlampaui atau akun terkunci sementara.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            rateLimited:
              value:
                success: false
                error: { code: RATE_LIMITED, message: Terlalu banyak request }
    InternalError:
      description: Kesalahan internal.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            internal:
              value:
                success: false
                error: { code: INTERNAL_ERROR, message: Terjadi kesalahan internal }
...
  schemas:
    UUID:
      type: string
      format: uuid
      example: 0193f406-8d2f-7d0c-bf34-61ff192e46de
...
    SuccessResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data: {}
        message: { type: string }
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string, example: Validasi gagal }
        fields:
          $ref: "#/components/schemas/ValidationFields"
    ValidationFields:
      type: object
      additionalProperties: { type: string }
      example: { email: Format email tidak valid }
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, const: false }
        error: { $ref: "#/components/schemas/ErrorObject" }
...
    Department:
      type: object
      required: [id, nama]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        nama: { type: string, maxLength: 100 }
    Position:
      type: object
      required: [id, department_id, nama]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        department_id: { $ref: "#/components/schemas/UUID" }
        nama: { type: string, maxLength: 100 }
...
    DepartmentListResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data: { type: array, items: { $ref: "#/components/schemas/Department" } }
    PositionListResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data: { type: array, items: { $ref: "#/components/schemas/Position" } }
...
    PhotoUploadRequest:
      type: object
      required: [foto]
      properties:
        foto:
          type: string
          format: binary
          description: JPG atau PNG; maksimal 5 MB.
    PhotoUpdateData:
      type: object
      required: [foto_profil_url]
      properties:
        foto_profil_url: { type: string, format: uri }
    PhotoUpdateResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/PhotoUpdateData" }
...